requirements.txt
bs4
flask
requests
python-dotenv
playwright
mysql-connector-python